refactor: 초대 링크 생성 기능 리팩토링
