+++ src/main/java/com/devnity/devnity/domain/admin/controller/AdminGenerationController.java
@@ -1,8 +1,8 @@
 package com.devnity.devnity.domain.admin.controller;
 
 import com.devnity.devnity.common.api.ApiResponse;
-import com.devnity.devnity.domain.admin.controller.dto.GenerationRequest;
-import com.devnity.devnity.domain.admin.controller.dto.GenerationResponse;
+import com.devnity.devnity.domain.admin.dto.request.GenerationRequest;
+import com.devnity.devnity.domain.admin.dto.response.GenerationResponse;
 import com.devnity.devnity.domain.admin.service.AdminGenerationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -16,26 +16,26 @@ import java.util.Map;
 @RequestMapping("/api/v1/admin/generations")
 public class AdminGenerationController {
 
-    private final AdminGenerationService service;
+  private final AdminGenerationService service;
 
-    @PostMapping
-    public ApiResponse<Boolean> create(@RequestBody GenerationRequest req) {
-        return ApiResponse.ok(service.create(req));
-    }
+  @PostMapping
+  public ApiResponse<Boolean> create(@RequestBody GenerationRequest req) {
+    return ApiResponse.ok(service.create(req));
+  }
 
-    @PutMapping("/{generationId}")
-    public ApiResponse<Boolean> update(@PathVariable Long generationId, @RequestBody GenerationRequest req) {
-        req.setId(generationId);
-        return ApiResponse.ok(service.update(req));
-    }
+  @PutMapping("/{generationId}")
+  public ApiResponse<Boolean> update(@PathVariable Long generationId, @RequestBody GenerationRequest req) {
+    req.setId(generationId);
+    return ApiResponse.ok(service.update(req));
+  }
 
-    @GetMapping
-    public ApiResponse<Map<String, List<GenerationResponse>>> get() {
-        return ApiResponse.ok(Collections.singletonMap("generations", service.getAll()));
-    }
+  @GetMapping
+  public ApiResponse<Map<String, List<GenerationResponse>>> get() {
+    return ApiResponse.ok(Collections.singletonMap("generations", service.getAll()));
+  }
 
-    @DeleteMapping("/{generationId}")
-    public ApiResponse<Boolean> delete(@PathVariable Long generationId) {
-        return ApiResponse.ok(service.delete(generationId));
-    }
+  @DeleteMapping("/{generationId}")
+  public ApiResponse<Boolean> delete(@PathVariable Long generationId) {
+    return ApiResponse.ok(service.delete(generationId));
+  }
 }
