--- src/main/java/com/devnity/devnity/domain/admin/controller/dto/GenerationResponse.java
@@ -1,19 +1,0 @@
-package com.devnity.devnity.domain.admin.controller.dto;
-
-import com.devnity.devnity.domain.user.entity.Generation;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class GenerationResponse {
-
-    Long generationId;
-    int sequence;
-
-    public static GenerationResponse from(Generation generation) {
-        return new GenerationResponse(generation.getId(), generation.getSequence());
-    }
-}
