+++ src/main/java/com/devnity/devnity/domain/admin/dto/request/GenerationRequest.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.domain.admin.controller.dto;
+package com.devnity.devnity.domain.admin.dto.request;
 
 import com.devnity.devnity.domain.user.entity.Generation;
 import lombok.AllArgsConstructor;
@@ -10,11 +10,15 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class GenerationRequest {
 
-    Long id;
-    int sequence;
+  Long id;
+  int sequence;
 
-    public GenerationRequest(int sequence) { this.sequence = sequence; }
+  public GenerationRequest(int sequence) {
+    this.sequence = sequence;
+  }
 
-    public Generation to() { return new Generation(sequence); }
+  public Generation to() {
+    return new Generation(sequence);
+  }
 
 }
