+++ src/main/java/com/devnity/devnity/domain/admin/dto/request/InvitationRequest.java
@@ -1,0 +1,34 @@
+package com.devnity.devnity.domain.admin.dto.request;
+
+import com.devnity.devnity.domain.user.entity.UserRole;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDate;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor
+@Builder
+@Getter
+public class InvitationRequest {
+
+  @NotBlank
+  private String course;
+
+  @NotNull
+  private Integer generation;
+
+  @NotBlank
+  private UserRole role;
+
+  @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
+  private LocalDate deadline;
+
+}
