+++ src/main/java/com/devnity/devnity/domain/admin/dto/response/CourseResponse.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.domain.admin.controller.dto;
+package com.devnity.devnity.domain.admin.dto.response;
 
 import com.devnity.devnity.domain.user.entity.Course;
 import lombok.AllArgsConstructor;
@@ -9,10 +9,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CourseResponse {
-   Long id;
-   String name;
 
-   public static CourseResponse from(Course course) {
-      return new CourseResponse(course.getId(), course.getName());
-   }
+  Long id;
+  String name;
+
+  public static CourseResponse from(Course course) {
+    return new CourseResponse(course.getId(), course.getName());
+  }
 }
