+++ src/main/java/com/devnity/devnity/domain/admin/dto/response/GenerationResponse.java
@@ -1,0 +1,19 @@
+package com.devnity.devnity.domain.admin.dto.response;
+
+import com.devnity.devnity.domain.user.entity.Generation;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class GenerationResponse {
+
+  Long generationId;
+  int sequence;
+
+  public static GenerationResponse from(Generation generation) {
+    return new GenerationResponse(generation.getId(), generation.getSequence());
+  }
+}
